+++ apps/lootables.js
@@ -115,7 +115,7 @@ export class LootablesApp extends Application {
                 newActorData = {
                     'flags': {
                         'core': {
-                            'sheetClass': 'dnd5e.LootSheet5eNPC'
+                            'sheetClass': 'dnd5e.LootSheetNPC5e'
                         },
                         'lootsheetnpc5e': {
                             'lootsheettype': 'Loot'