+++ apps/savingthrow.js
@@ -327,8 +327,8 @@ export class SavingThrow {
     static async _rollAbility(data, request, requesttype, rollmode, ffwd, e) {
         //let actor = game.actors.get(data.actorid);
         let tokenOrActor = await fromUuid(data.uuid);
-        let actor = tokenOrActor.actor ? tokenOrActor.actor : tokenOrActor;
-        let fastForward = ffwd || (e && (e.shiftKey || e.altKey || e.ctrlKey || e.metaKey));
+        let actor = tokenOrActor?.actor ? tokenOrActor.actor : tokenOrActor;
+        let fastForward = ffwd || (e.shiftKey || e.altKey || e.ctrlKey || e.metaKey);
 
         if (actor != undefined) {
             if (requesttype == 'dice') {
@@ -366,10 +366,20 @@ export class SavingThrow {
             if (msgtoken != undefined && msgtoken.roll == undefined) {
                 //let actor = game.actors.get(msgtoken.actorid);
                 let tokenOrActor = await fromUuid(msgtoken.uuid);
-                let actor = tokenOrActor.actor ? tokenOrActor.actor : tokenOrActor;
+                let actor = tokenOrActor?.actor ? tokenOrActor.actor : tokenOrActor;
                 if (actor != undefined) {
                     //roll the dice, using standard details from actor
-                    let e = Object.assign({}, evt, msgtoken.keys);
+                    let keys = msgtoken.keys || {};
+                    let e = Object.assign({}, evt);
+                    e.ctrlKey = evt.ctrlKey;
+                    e.altKey = evt.altKey;
+                    e.shiftKey = evt.shiftKey;
+                    e.metaKey = evt.metaKey;
+
+                    for (let [k, v] of Object.entries(keys))
+                        e[k] = evt[k] || v;
+                    MonksTokenBar.system.parseKeys(e, keys);
+
                     promises.push(SavingThrow._rollAbility({ id: id, uuid: msgtoken.uuid }, request, requesttype, rollmode, fastForward, e));
                 }
             }
@@ -454,7 +464,7 @@ export class SavingThrow {
                 } else if (update.error === true) {
                     //let actor = game.actors.get(msgtoken.actorid);
                     let tokenOrActor = await fromUuid(msgtoken.uuid);
-                    let actor = tokenOrActor.actor ? tokenOrActor.actor : tokenOrActor;
+                    let actor = tokenOrActor?.actor ? tokenOrActor.actor : tokenOrActor;
 
                     ui.notifications.warn(msgtoken.name + ': ' + update.msg);
 
@@ -750,11 +760,11 @@ Hooks.on("renderChatMessage", async (message, html, data) => {
             if (msgtoken) {
                 //let actor = game.actors.get(msgtoken.actorid);
                 let tokenOrActor = await fromUuid(msgtoken.uuid);
-                let actor = tokenOrActor.actor ? tokenOrActor.actor : tokenOrActor;
+                let actor = tokenOrActor?.actor ? tokenOrActor.actor : tokenOrActor;
 
                 $(item).toggle(game.user.isGM || rollmode == 'roll' || rollmode == 'gmroll' || (rollmode == 'blindroll' && actor.isOwner));
 
-                if (game.user.isGM || actor.isOwner)
+                if (game.user.isGM || actor?.isOwner)
                     $('.item-image', item).on('click', $.proxy(SavingThrow._onClickToken, this, msgtoken.id))
                 $('.item-roll', item).toggle(msgtoken.roll == undefined && (game.user.isGM || (actor.isOwner && rollmode != 'selfroll'))).click($.proxy(SavingThrow.onRollAbility, this, msgtoken.id, message, false));
                 $('.dice-total', item).toggle((msgtoken.error === true || msgtoken.roll != undefined) && (game.user.isGM || rollmode == 'roll' || (actor.isOwner && rollmode != 'selfroll')));